# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Security scanning platform — manages targets, scan jobs, reports, and triage rules.",
        "title": "NexusSec API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/scans/{id}/triage": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Lấy danh sách triage rules của một scan (mà user sở hữu).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "triage"
                ],
                "summary": "Get triage rules",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Scan Job ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Danh sách triage rules",
                        "schema": {
                            "$ref": "#/definitions/github_com_nexussec_nexussec_pkg_response.JSONResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_nexussec_nexussec_pkg_response.JSONResponse"
                        }
                    },
                    "404": {
                        "description": "Scan không tồn tại",
                        "schema": {
                            "$ref": "#/definitions/github_com_nexussec_nexussec_pkg_response.JSONResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_nexussec_nexussec_pkg_response.JSONResponse"
                        }
                    }
                }
            }
        },
        "/scans/{id}/triage/{fingerprint}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Tạo hoặc cập nhật triage rule cho một lỗ hổng (theo fingerprint).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "triage"
                ],
                "summary": "Upsert triage rule",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Scan Job ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Fingerprint của vulnerability",
                        "name": "fingerprint",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Triage rule data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_nexussec_nexussec_internal_domain_model.TriageRuleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cập nhật thành công",
                        "schema": {
                            "$ref": "#/definitions/github_com_nexussec_nexussec_pkg_response.JSONResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_nexussec_nexussec_pkg_response.JSONResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_nexussec_nexussec_pkg_response.JSONResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_nexussec_nexussec_pkg_response.JSONResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_nexussec_nexussec_pkg_response.JSONResponse"
                        }
                    }
                }
            }
        },
        "/targets": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Lấy danh sách tất cả target của user hiện tại.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "targets"
                ],
                "summary": "List scan targets",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_nexussec_nexussec_pkg_response.JSONResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/internal_gateway_handler.targetResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_nexussec_nexussec_pkg_response.JSONResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_nexussec_nexussec_pkg_response.JSONResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Tạo một target mới để scan. URL được kiểm tra SSRF. Trả về 409 nếu URL đã tồn tại.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "targets"
                ],
                "summary": "Create a scan target",
                "parameters": [
                    {
                        "description": "Thông tin target",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_gateway_handler.createTargetRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Tạo thành công",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_nexussec_nexussec_pkg_response.JSONResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_gateway_handler.targetResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "URL không hợp lệ hoặc SSRF blocked",
                        "schema": {
                            "$ref": "#/definitions/github_com_nexussec_nexussec_pkg_response.JSONResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_nexussec_nexussec_pkg_response.JSONResponse"
                        }
                    },
                    "409": {
                        "description": "Target URL đã tồn tại",
                        "schema": {
                            "$ref": "#/definitions/github_com_nexussec_nexussec_pkg_response.JSONResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_nexussec_nexussec_pkg_response.JSONResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_nexussec_nexussec_internal_domain_model.TriageRuleRequest": {
            "type": "object",
            "properties": {
                "is_false_positive": {
                    "type": "boolean"
                },
                "is_muted": {
                    "type": "boolean"
                },
                "notes": {
                    "type": "string"
                }
            }
        },
        "github_com_nexussec_nexussec_pkg_response.JSONResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "internal_gateway_handler.createTargetRequest": {
            "type": "object",
            "required": [
                "base_url",
                "name"
            ],
            "properties": {
                "base_url": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 1
                }
            }
        },
...
        "internal_gateway_handler.targetResponse": {
            "type": "object",
            "properties": {
                "base_url": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Nhập: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}